+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### v0.1.4 - 2026-04-29
+
+- Fix reading concatenated gzip files. 
+
+### v0.1.3 - 2026-04-28
+
+- Improve parsing speed.
+
 ### v0.1.2 - 2026-04-27
 
 - Fix a bug of writer, which did not flush data in the end.